updated README

+++ submit/lab7-sol/multiple-defs/f.c
@@ -1,5 +1,5 @@
-extern double x;
+extern int x;
 
 void f() {
-  x = -0.0;
+  x = -0;
 }
